refactor(api/test): use TestApp CollectionFixture
