+++ modules/back-end/tests/Application.IntegrationTests/TestAppCollection.cs
@@ -1,0 +1,9 @@
+namespace Application.IntegrationTests;
+
+[CollectionDefinition(nameof(TestApp))]
+public class TestAppCollection : ICollectionFixture<TestApp>
+{
+    // This class has no code, and is never created. Its purpose is simply
+    // to be the place to apply [CollectionDefinition] and all the
+    // ICollectionFixture<> interfaces.
+}
